Fix: resolve eslint errors
